pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sommmelier"
version = "0.1.0"
description = "AI-driven Marketing Mix Modeling powered by Google Meridian"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11,<3.13"
authors = [
    { name = "Giorgio Liapakis" }
]
keywords = ["mmm", "marketing-mix-model", "meridian", "marketing", "attribution", "ai", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "google-meridian>=1.4.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "numpy>=2.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "modal>=0.64.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
gpu = [
    "google-meridian[and-cuda]>=1.0.0",
]

[project.scripts]
sommmelier = "mmm.cli.main:app"

[project.urls]
Homepage = "https://github.com/giorgioliapakis/sommmelier"
Repository = "https://github.com/giorgioliapakis/sommmelier"

[tool.hatch.build.targets.wheel]
packages = ["mmm"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=mmm"
